refactor(rom-browser): split RomBrowserApp impl into focused submodules (#2828)
Distribute the ~2,530-line impl RomBrowserApp and the ApplicationHandler
impl across focused submodules under rom_browser/app/: catalog_state,
textures, input, actions, handler, and a render/ subdir (mod + grid +
search + filter + detail). mod.rs retains the type definitions,
associated consts, new()/run(), and module wiring. Methods called across
module boundaries are elevated to pub(super); the external API
(RomBrowserApp::new/run, BrowserResult) is unchanged.

Pure code move: no behavior change. All 38 ROM-browser tests pass; every
impl file is now under ~800 lines (largest: render/mod.rs at 601).
